requirements.txt
streamlit
requests
python-dotenv
yagmail
schedule
openai
